Support serving UI from sub path
