+++ config/base.yaml
@@ -1,3 +1,3 @@
 temporalGrpcAddress: 127.0.0.1:7233
 port: 8080
-uiRootPath: /
+
