+++ config/development.yaml
@@ -1,3 +1,4 @@
+publicPath:
 enableUi: true
 enableOpenApi: true
 cors:
