+++ server/route/openapi.go
@@ -20,15 +20,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package routes
+package route
 
 import (
-	"github.com/labstack/echo/v4"
+	"bytes"
+	"embed"
+	"io/fs"
 	"net/http"
+
+	"github.com/labstack/echo/v4"
 )
 
-func SetHealthRoute(e *echo.Echo) {
-	e.GET("/healthz", func(c echo.Context) error {
-		return c.JSON(http.StatusOK, map[string]string{"status": "OK"})
-	})
+// SetAPIRoutes sets api routes
+func SetSwaggerUIRoutes(e *echo.Group, indexHTML []byte, assets embed.FS) {
+	e.GET("/openapi", buildSwaggerUIHandler(indexHTML))
+	e.GET("/openapi/*", buildSwaggerUIAssetsHander(assets))
+}
+
+func buildSwaggerUIHandler(indexHTML []byte) echo.HandlerFunc {
+	return func(c echo.Context) (err error) {
+		return c.Stream(200, "text/html", bytes.NewBuffer(indexHTML))
+	}
+}
+
+func buildSwaggerUIAssetsHander(assets embed.FS) echo.HandlerFunc {
+	stream := fs.FS(assets)
+	stream, _ = fs.Sub(stream, "generated")
+	handler := http.FileServer(http.FS(stream))
+	return echo.WrapHandler(handler)
 }
